requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
cffi==1.15.1
charset-normalizer==3.1.0
discord==2.2.2
discord.py==2.2.2
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
nextcord==2.4.2
pycparser==2.21
PyNaCl==1.4.0
python-dotenv==1.0.0
six==1.16.0
typing_extensions==4.5.0
wavelink @ git+https://github.com/Matezcp/Wavelink.git@5255abf9e1f68840c68160f873b1019cb65d2bd1
yarl==1.9.2
